fix: supporterPlusLegal description shared between paymentTs&Cs and ContributionOrderSummary
